cr: support setAttr(mtime) for directories
Previously, we treated all setAttr(mtime) operations as if they
happened on files, when in fact that is a valid operation for a
directory. Even though setAttrs affect directory entries, they really
apply to the target node itself (which may be moved around in another
branch).  Thus, the CR chains "pin" setAttr operations to the node,
and then apply them to the node's final merged parent block.

This is somewhat straightforward with files, since all file-related
operations work this way (syncOps+setAttrOps).  With directories
however, this isn't the case, since a directory may both have its
mtime set and have a new entry created within it; therefore, we must
pick and choose which operations get applied to the parent block
containing the directory's directory entry.

To make this happen, this commit does the following:

* A crChain identifies a file type by fetching the actual block, when
  it isn't possible to identify it using operations alone.

* ConflictResolver splits up the actions relating to a directory's
  mtime into a separate list, and applied them to the directory's
  parent.  If the parent didn't yet have any actions, it must be added
  to the list of unmergedPaths.

* Add a new error type for when a block can't be decoded into its
  expected type, so we can easily distinguish that and use it to help
  determine a chain's type.

* If a directory's mtime is modified in both branches, resolve the
  conflict by creating a symlink (imbued with the unmerged mtime)
  pointing to the real entry.

* In order to support reading the mtime of a symlink, change the DSL
  to not resolve the final symlink in a path for the mtime operations,
  and fix up the libkbfs engine accordingly.

Issue: KBFS-1163